refactor(filid): clear zero-peer-file warnings in hooks
Two different causes, two different fixes.

`*.entry.ts` (4 hook fractals): these are the esbuild entry points that
src/hooks/INTENT.md documents as the hook convention — stdin -> handler
-> stdout, no logic. They belong at the fractal root, which is exactly
the case `additional-allowed` exists for; the rule's own suggestion says
so. Registered path-scoped, so the allowance cannot leak to other dirs.
Only hooks/setup was in this branch's review scope — the other three
carried the same warning unreported.

`hooks/shared` (5 predicate peers): not a convention, just implementation
sitting at the root. Every sibling fractal (boundaryDetector,
changeTracker, intentInjector, preToolValidator) keeps logic in an organ
behind a facade; shared/ was the outlier. Moved to shared/utils/ and left
shared.ts re-exporting them by concrete path (never a barrel — hook
bundles must not pull one in). Nothing outside shared/ imported these
files directly, so no consumer changes.

structure_validate over src/hooks: 6 violations -> 0 (84 checks pass).
Hook bundle guards pass and every bundle is byte-identical to before the
move (pre-tool-use 23718, user-prompt-submit 11227, agent-enforcer 9393,
setup 23873) — esbuild inlines, so relocation costs nothing. 1124 tests
pass, typecheck clean.
